requirements.txt
## requirements.txt
```
requests>=2.25.0
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4.0
hvac>=0.11.0
fhir.resources>=6.0.0  # For FHIR resource parsing
hcuppy>=0.0.5  # For Elixhauser scoring
```